requirements.txt
matplotlib
windrose
numpy
kafka-python